Rename "atomics" to "atoms"
