+++ lib/components/atoms/cards/member-card/member-card.type.ts
