+++ lib/components/atoms/cards/project-cards/new-project-card/new-project-card.tsx
