+++ lib/components/atoms/cards/space-card/space-card.tsx
