+++ lib/components/atoms/hamburger/hamburger.type.ts
